+++ src/config/eslintrc.js
@@ -1,3 +1,9 @@
+const { hasPkgDep } = require("../utils");
+
 module.exports = {
-  extends: ["@culturehq"]
+  extends: [
+    "@culturehq",
+    hasPkgDep("react") && "@culturehq/eslint-config/react",
+    hasPkgDep("typescript") && "@culturehq/eslint-config/typescript"
+  ].filter(Boolean)
 };